[fix]helperが動いた
