+++ Assets/TitleScenes/Scripts/JsonHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47de97bdbfb4abd4b822db9bc573ce54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
